refactor: dictionary
