+++ drama-queen/src/i18n/dictionary.ts
@@ -19,13 +19,3 @@ export const dictionary = {
 
 export type Dictionary = typeof dictionary
 export type DictionaryLang = 'fr' | 'en'
-
-export type Entries<T> = {
-  [K in keyof T]: [K, T[K]]
-}[keyof T][]
-
-export type AbstractDictionary<T> = {
-  [Key in keyof T]: {
-    [Lang in DictionaryLang]: string
-  }
-}
